+++ sources/packages/backend/apps/queue-consumers/src/queue-consumers.module.ts
@@ -4,12 +4,12 @@ import { QueueModule } from "@sims/services/queue";
 import { StartApplicationAssessmentProcessor } from "./processors";
 import { WorkflowClientService, ZeebeModule } from "@sims/services";
 import { DatabaseModule } from "@sims/sims-db";
-import { IER12IntegrationService } from "@sims/integrations/institution-integration/ier-integration";
+import { IER12IntegrationService } from "@sims/integrations/institution-integration/ier12-integration";
 import {
   SshService,
   StudentAssessmentService,
 } from "@sims/integrations/services";
-import { IER12IntegrationModule } from "@sims/integrations/institution-integration/ier-integration/ier12-integration.module";
+import { IER12IntegrationModule } from "@sims/integrations/institution-integration/ier12-integration/ier12-integration.module";
 import { IER12IntegrationScheduler } from "./processors/schedulers/institution-integration/ier12-Integrations/ier12-integration.scheduler";
 
 @Module({
